Use PROMPT_DIRTRIM instead of _short_wdir().
The latter stopped working at some point in the sense that it was messing up
the prompt so that bash could not determine its actual real length in
characters any more, thus completely breaking the new-line-wrapping.

Plus, there is of course no need to re-invent the wheel if there is already
such a function built-in.